services/flask-oidc-proxy/requirements.txt
Flask==3.1.0
Flask-Session==0.8.0
requests==2.32.3
PyJWT[crypto]==2.10.1
cryptography==44.0.0
redis==5.2.1
gunicorn==23.0.0
